% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivocal.R
\name{nivocal-shiny}
\alias{nivocal-shiny}
\alias{nivocalOutput}
\alias{renderNivocal}
\alias{nivocal_html}
\title{Shiny bindings for nivocal}
\usage{
nivocalOutput(outputId, width = "100\%\%", height = "400px")

renderNivocal(expr, env = parent.frame(), quoted = FALSE)

nivocal_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a nivocal}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using nivocal within Shiny
applications and interactive Rmd documents.
}
